Refactor separate components for Editing and Dashboards
