+++ src/ControlBarContainer/ControlBarContainer.js
@@ -1,24 +1,13 @@
-import React, { Fragment } from 'react';
+import React from 'react';
 import { connect } from 'react-redux';
 
-import ControlBar from 'd2-ui/lib/controlbar/ControlBar';
-import SvgIcon from 'd2-ui/lib/svg-icon/SvgIcon';
-import Chip from 'd2-ui/lib/chip/Chip';
-
-import D2IconButton from '../widgets/D2IconButton';
-import Filter from './Filter';
-
-import * as fromActions from '../actions';
-import * as fromReducers from '../reducers';
-import { orObject } from '../util';
-import { blue800 } from '../../../d2-ui/node_modules/material-ui/styles/colors';
+import EditBar from './EditBar';
+import DashboardsBar from './DashboardsBar';
+import { fromSelected } from '../reducers';
 
 import './ControlBarContainer.css';
 
-const styles = {
-    scrollWrapper: {
-        padding: '10px 6px 0 6px',
-    },
+const style = {
     leftControls: {
         display: 'inline-block',
         fontSize: 16,
@@ -29,183 +18,25 @@ const styles = {
         display: 'inline-block',
         float: 'right',
     },
-    expandButtonWrap: {
-        textAlign: 'center',
-    },
 };
 
-export const CONTROL_BAR_ROW_HEIGHT = 36;
-const EXPANDED_ROW_COUNT = 10;
-const CONTROL_BAR_OUTER_HEIGHT_DIFF = 24;
-
-const getInnerHeight = (isExpanded, rows) =>
-    (isExpanded ? EXPANDED_ROW_COUNT : rows) * CONTROL_BAR_ROW_HEIGHT;
-
-const getOuterHeight = (isExpanded, rows) =>
-    getInnerHeight(isExpanded, rows) + CONTROL_BAR_OUTER_HEIGHT_DIFF;
-
-const onDashboardSelectWrapper = (id, onClick) => () => onClick(id);
-
-const ControlBarComponent = ({
-    dashboards,
-    name,
-    edit,
-    rows,
-    expandable,
-    isExpanded,
-    onChangeHeight,
-    onToggleExpanded,
-    onNewClick,
-    onChangeFilterName,
-    onDashboardSelect,
-}) => {
-    const contentWrapperStyle = Object.assign(
-        {},
-        styles.scrollWrapper,
-        { overflowY: isExpanded ? 'auto' : 'hidden' },
-        { height: getInnerHeight(isExpanded, rows) }
-    );
-
-    const controlBarHeight = getOuterHeight(
-        edit ? false : isExpanded,
-        edit ? 1 : rows
-    );
-
-    return (
-        <ControlBar
-            height={controlBarHeight}
-            onChangeHeight={onChangeHeight}
-            editMode={edit}
-            expandable={expandable}
-        >
-            <div style={contentWrapperStyle}>
-                <div style={styles.leftControls}>
-                    {edit ? (
-                        <div>Save changes</div>
-                    ) : (
-                        <Fragment>
-                            <D2IconButton
-                                style={{
-                                    width: 36,
-                                    height: 36,
-                                    marginRight: 10,
-                                }}
-                                onClick={onNewClick}
-                            />
-                            <Filter
-                                name={name}
-                                onChangeName={onChangeFilterName}
-                            />
-                        </Fragment>
-                    )}
-                </div>
-                <div style={styles.rightControls}>
-                    {edit ? (
-                        <div>Exit without saving</div>
-                    ) : (
-                        <div
-                            style={{
-                                position: 'relative',
-                                top: '6px',
-                                left: '-10px',
-                                cursor: 'pointer',
-                            }}
-                            onClick={() => alert('show list view')}
-                        >
-                            <SvgIcon icon="List" />
-                        </div>
-                    )}
-                </div>
-                {!edit &&
-                    dashboards.map(dashboard => (
-                        <Chip
-                            key={dashboard.id}
-                            label={dashboard.name}
-                            avatar={dashboard.starred ? 'star' : null}
-                            onClick={onDashboardSelectWrapper(
-                                dashboard.id,
-                                onDashboardSelect
-                            )}
-                        />
-                    ))}
-            </div>
-            <div style={styles.expandButtonWrap}>
-                <div
-                    onClick={onToggleExpanded}
-                    style={{
-                        paddingTop: 4,
-                        fontSize: 11,
-                        fontWeight: 700,
-                        color: blue800,
-                        textTransform: 'uppercase',
-                        cursor: 'pointer',
-                        visibility: edit ? 'hidden' : 'visible',
-                    }}
-                >
-                    {isExpanded ? 'Show less' : 'Show more'}
-                </div>
-            </div>
-        </ControlBar>
+const ControlBar = ({ edit }) => {
+    return edit ? (
+        <EditBar style={style} />
+    ) : (
+        <DashboardsBar controlsStyle={style} />
     );
 };
 
 const mapStateToProps = state => {
-    const { fromDashboards, fromSelected, fromFilter } = fromReducers;
-
     return {
-        dashboards: fromDashboards.sGetFromState(state),
-        name: fromFilter.sGetFilterName(state),
         edit: fromSelected.sGetSelectedEdit(state),
-        rows: (state.controlBar && state.controlBar.rows) || 1,
-        isExpanded:
-            state.controlBar.expanded &&
-            state.controlBar.rows < EXPANDED_ROW_COUNT,
     };
 };
 
-const mergeProps = (stateProps, dispatchProps, ownProps) => {
-    const { dashboards, name, edit, rows, isExpanded } = stateProps;
-    const { dispatch } = dispatchProps;
-    const { fromControlBar, fromFilter, fromSelected } = fromActions;
+const ControlBarCt = connect(mapStateToProps, null)(ControlBar);
 
-    const filteredDashboards = Object.values(orObject(dashboards)).filter(
-        d => d.name.toLowerCase().indexOf(name) !== -1
-    );
+export default ControlBarCt;
 
-    return {
-        ...stateProps,
-        ...ownProps,
-        dashboards: [
-            ...filteredDashboards.filter(d => d.starred),
-            ...filteredDashboards.filter(d => !d.starred),
-        ],
-        expandable: !isExpanded && !edit,
-        onChangeHeight: (newHeight, onEndDrag) => {
-            const newRows = Math.max(
-                1,
-                Math.floor((newHeight - 24) / CONTROL_BAR_ROW_HEIGHT)
-            );
-
-            if (newRows !== rows) {
-                dispatch(
-                    fromControlBar.acSetControlBarRows(
-                        Math.min(newRows, EXPANDED_ROW_COUNT)
-                    )
-                );
-            }
-        },
-        onToggleExpanded: () => {
-            dispatch(fromControlBar.acSetControlBarExpanded(!isExpanded));
-        },
-        onNewClick: () => dispatch(fromSelected.tNewDashboard()),
-        onChangeFilterName: name => dispatch(fromFilter.acSetFilterName(name)),
-        onDashboardSelect: id =>
-            dispatch(fromSelected.tSetSelectedDashboardById(id)),
-    };
-};
-
-export const ControlBarContainer = connect(mapStateToProps, null, mergeProps)(
-    ControlBarComponent
-);
-
-export default ControlBarContainer;
+export const CONTROL_BAR_ROW_HEIGHT = 36;
+export const CONTROL_BAR_OUTER_HEIGHT_DIFF = 24;
